Add basic sensitivity functionality
Sensitivities are saved to the binary file as a 1D array. This must be
reshaped into a n_vars,n_sensitivity_params array upon reading.
Possible gotcha:
The variable n_vars is set currently to n_species + 3 because this seems
to be what Cantera sets netw.n_vars to after initializing the network.
The network is not initialized until after the first time step, so the
first row of outArray depends on n_vars being set to the correct number