Removed redirection of stdout to devnull for case setup in multi.
During multi run, stdout was redirected to stdout to avoid printing
input keywords for every case. However, this had the unintended side
effect of also preventing the display of any syntax errors or omitted
keywords in the input file. So, this redirection has been removed
in order to see any input file errors.